requirements.txt
openai==1.30.1
pandas==2.2.2
plotly==5.22.0
python-dotenv==1.0.1
python_bcrypt==0.3.2
streamlit==1.34.0
streamlit_authenticator==0.3.2
